requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
aniso8601==9.0.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
bcrypt==4.1.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.5.0
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
frozenlist==1.4.1
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.60.0
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
passlib==1.7.4
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.23
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
WTForms==3.1.1
yarl==1.9.4
